+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 | Version | Changes                                                    |
 | --------|------------------------------------------------------------|
 | 0.0.3   | Template variables support                                 |
+|         | `Uniques` format support                                   |
 | 0.0.2   | Grafana Toolkit implemented                                |  
 | 0.0.1   | First working version                                      |
 |         | Newrelic Insights datasource                               |